refactor: add command module
